updated stylessheets
