+++ findingRides/App.tsx
@@ -15,7 +15,7 @@ export default function App() {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: "FFFFFF",
+    backgroundColor: "#5D737E",
     alignItems: "center",
     justifyContent: "center",
   },
