+++ findingRides/src/screens/HomeStyles.ts
@@ -1,0 +1,15 @@
+import { ViewStyle } from "react-native";
+
+export interface Styles {
+  container: ViewStyle;
+}
+
+export const HomeStyles = () => {
+  return {
+    container: {
+      flex: 1,
+      alignItems: "center",
+      justifyContent: "center",
+    },
+  };
+};
